[web] Move JS content to its own `.js` files (#117691)
